+++ app/templates/test/policies/isUser.js
@@ -1,5 +1,5 @@
 var assert = require('assert'),
-    isUser = require('../../api/policies/isUser');
+    isUser = require('../../api/policies/isAuthenticated');
 
 describe("policies:isUser", function () {
     it("TODO: write this test", function () {