+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Improvement: Rewrite default blueprints for CRUD and simplify them;
 - Improvement: Returns meta information for `GET /v1/:model` in body and `Content-Range` in headers;
 - Improvement: Add questions for Facebook and Twitter tokens;
+- Improvement: Rename `isUser` policy to `isAuthenticated`, it's more appropriate name;
 - Fix: Getting status code from passport in `AuthController`;
 
 ## Version 0.3.0